backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
scikit-surprise==1.1.4
google-api-python-client==2.149.0
python-dotenv==1.0.1
matplotlib==3.9.2
tensorflow==2.17.0
werkzeug==3.0.4
setuptools==75.1.0
seaborn==0.13.2